requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
psycopg2>=2.6,<2.9.9
drf-spectacular>=0.25,<=0.27.1